create a folder architecture for a clean install of wordpress.
create a folder architecture to have a clean wordpress installation.
first, create a public folder to contain wordpress. Inside it create a wp
folder to have the cms and a wp-content for themes, plugins and specific
content.
create a dedicated index.php to bootstrap the new architecture.
create a wp-config which import a specific file with configurations for
your wordpress instance like credentials for the database.
this configuration file is not into the public folder, it's just for
respect the good practices.
the configuration file has to be in the root folder, an example of local
configuration named local-config.sample.php was added.