create a docker configuration to run wordpress into containers.
create a dockerfile to have php 7.x, composer  and mysqli with a user with
UUID and GID 1000 (change it to have the same as the current user on
host).
create a docker compose configuration.
the compose run mariadb with a dedicated
docker volume to save configurations and other stuff from our wordpress
instance.
it run a wordpress container which is connected to the db and run
wordpress with the php developer server.